(DOCSP-11948): Move StatusBadge to its own module
